requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.3
debugpy==1.4.1
decorator==5.0.9
discord.py==1.7.1
entrypoints==0.3
google-api-core==1.31.5
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.1
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.1
jupyter-core==4.7.1
matplotlib-inline==0.1.2
multidict==5.1.0
nest-asyncio==1.5.1
packaging==21.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.20
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-resize-image==1.1.20
pytz==2021.3
pywin32==228
pyzmq==22.2.1
requests==2.27.0
rsa==4.8
six==1.16.0
termcolor==1.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.7
wcwidth==0.2.5
yarl==1.6.3